Add files via upload
J.A.R.V.I.S. is an AI that functions as Tony Stark's assistant, running and taking care of all the internal systems of Stark's buildings and the Iron Man suits. In Peter David's novelization of Iron Man, J.A.R.V.I.S. is said to be an acronym for "Just a Rather Very Intelligent System".